requirements.txt
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
optuna>=3.0.0
websockets>=12.0
aiohttp>=3.9.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.66.0
pyarrow>=14.0.0
